list products

+++ src/modules/Products/useCases/deleteProductInInventory/DeleteProductUseCase.ts
@@ -1,5 +1,6 @@
 import { Inventory } from "@modules/Products/infra/entities/Inventory";
 import { IInventoryRepository } from "@modules/Products/IRepositories/IInventoryRepository";
+import { AppError } from "@shared/errors/AppError";
 import { inject, injectable } from "tsyringe";
 
 @injectable()
@@ -10,6 +11,12 @@ export class DeleteProductUseCase {
   ) {}
 
   async execute(product_id: string): Promise<void> {
+    const productExists = await this.inventoryRepository.getById(product_id);
+
+    if (!productExists) {
+      throw new AppError("Product not exists");
+    }
+
     await this.inventoryRepository.deleteProduct(product_id);
   }
 }
